// apps/frontend/src/client/post-store/models/AiContentPostRequest.ts
/**
 * Post Store API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class AiContentPostRequest {
    'text': string;
    'actor': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "text",
            "baseName": "text",
            "type": "string",
            "format": ""
        },
        {
            "name": "actor",
            "baseName": "actor",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AiContentPostRequest.attributeTypeMap;
    }

    public constructor() {
    }
}