// apps/frontend/src/client/post-store/models/AiContentResponse.ts
/**
 * Post Store API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AiContentResponseRateLimit } from '../models/AiContentResponseRateLimit';
import { HttpFile } from '../http/http';

export class AiContentResponse {
    'text': string;
    'rateLimit': AiContentResponseRateLimit;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "text",
            "baseName": "text",
            "type": "string",
            "format": ""
        },
        {
            "name": "rateLimit",
            "baseName": "rateLimit",
            "type": "AiContentResponseRateLimit",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AiContentResponse.attributeTypeMap;
    }

    public constructor() {
    }
}